Convert image to be RGB when reading
This adapts BioFormatsImageServer to convert an 8-bit, 3-channel image with red, green and blue LUTs to become a packed-int RGB image.

The purpose of this is to make it compatible with color deconvolution and other QuPath RGB-specific commands. The cost is that changing LUTs and channel names is no longer possible.

It was inspired by the observation that the conversion mentioned in https://github.com/qupath/qupath/issues/904 gave an image that wasn't treated as RGB.