Enable setting openslide path as a system property
The aim was to make it possible to load libopenslide-jni.jnilib from an absolute path, before OpenSlideJNI tries to load it from a path stored in the properties.

Sadly, it didn't work. It's necessary to set the working directory to be the one containing libopenslide-jni.jnilib, or alternatively to set an openslide.jar that has a suitable property early in the classpath.