Add ROI.isSimplePolygon() method
This fundamentally changes some aspects of how the object hierarchy works... but hopefully in a positive way.

ROI.isSimplePolygon() should return true for single-piece polygons without holes or self-intersections. A detection with a simple polygon ROI is considered to be inside an annotation ROI if the annotation ROI contains the centroid.

*However* if the detection does *not* have a simple polygon ROI, then the more expensive covers-with-tolerance check is performed.

The intention is to avoid the problem of a ROI with holes or multiple pieces being assigned to an annotation that doesn't even contain it at all (because the centroid is outside).

However it *doesn't* address the problem of a C-shaped ROI, with a centroid outside the original ROI still causing an unexpected result.